SC-7804 same name for component and page might be cause for mem leak in build. renamed component
